Remove Analyse endpoint
